[refactoring] actually check if streams are live
