+++ legistream/tasks.py
@@ -1,7 +1,8 @@
 from __future__ import absolute_import, unicode_literals
 
 from celery import shared_task
-from . import statuscheck
+from .status_checker import statuscheck
+
 
 @shared_task
 def get_streams():
