+++ legistream_site/celery.py
@@ -5,6 +5,7 @@ from . import settings
 
 from celery import Celery
 
+
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'legistream_site.settings')
 
 app = Celery('legistream_site')
@@ -14,8 +15,7 @@ app.config_from_object('django.conf:settings', namespace='CELERY')
 app.conf.beat_schedule = {
     'write_streams_bg': {
         'task': 'legistream.tasks.get_streams',
-        'schedule': 60,
-
+        'schedule': 300,
     }
 }
